Dont cast expected bloom filter size to int
Patch by marcuse; reviewed by Jason Brown for CASSANDRA-9959

+++ src/java/org/apache/cassandra/db/compaction/CompactionManager.java
@@ -553,8 +553,8 @@ public class CompactionManager implements CompactionManagerMBean
 
             long totalkeysWritten = 0;
 
-            int expectedBloomFilterSize = Math.max(cfs.metadata.getIndexInterval(),
-                                                   (int) (SSTableReader.getApproximateKeyCount(sstableAsSet, cfs.metadata)));
+            long expectedBloomFilterSize = Math.max(cfs.metadata.getIndexInterval(),
+                                                    SSTableReader.getApproximateKeyCount(sstableAsSet, cfs.metadata));
             if (logger.isDebugEnabled())
                 logger.debug("Expected bloom filter size : " + expectedBloomFilterSize);
 
@@ -734,7 +734,7 @@ public class CompactionManager implements CompactionManagerMBean
 
     public static SSTableWriter createWriter(ColumnFamilyStore cfs,
                                              File compactionFileLocation,
-                                             int expectedBloomFilterSize,
+                                             long expectedBloomFilterSize,
                                              SSTableReader sstable)
     {
         FileUtils.createDirectory(compactionFileLocation);
